requirements.txt
pyyaml
cfn_flip
boto3